# api/openapi.yaml
openapi: 3.1.0
info:
  title: App Parceiros API
  version: 0.2.0
  description: >
    Contrato da API do App Parceiros. É a fonte da verdade: altere aqui
    antes de implementar e gere os clientes a partir deste arquivo.


    Erros vêm sempre como `{code, message}` (schema `Error`): `code` é
    estável e serve para o programa; `message` já vem em português, pronta
    para mostrar ao usuário. O status segue o tipo do erro: 400 para JSON
    ou parâmetro malformado (`invalid_json`, `invalid_request`), 401 sem
    login, 402 workspace suspenso, 403 sem permissão, 404 não encontrado,
    409 conflito, 410 link vencido, 422 dado inválido, 429 limite de
    chamadas, 502 provedor externo fora do ar, 503 recurso não configurado
    no servidor e 500 `internal_error`.


    As rotas de um workspace ficam sob `/v1/workspaces/{workspaceId}`. Quem
    não é membro recebe 404 `workspace_not_found`. Com o workspace suspenso
    por falta de pagamento, só continuam valendo ver o workspace, sair dele
    (ou remover afiliados) e a assinatura; o resto responde 402
    (`workspace_suspended`, ou `workspace_suspended_owner` para o dono).
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/auth/register:
    post:
      summary: Cria uma conta com e-mail e senha (provedor interno)
      description: >
        Só existe com AUTH_PROVIDER=internal. Cria a conta, manda o link de
        confirmação do e-mail e já abre uma sessão. O usuário e o workspace
        pessoal são criados no primeiro acesso autenticado (ex.: GET /v1/me),
        como em qualquer provedor.
      operationId: authRegister
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthRegister" }
      responses:
        "201":
          description: Conta criada e sessão aberta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "409":
          description: "`email_taken`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`invalid_email`, `invalid_name` ou `weak_password`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: "`too_many_attempts`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/login:
    post:
      summary: Entra com e-mail e senha (provedor interno)
      operationId: authLogin
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthLogin" }
      responses:
        "200":
          description: Sessão aberta. Use o token como Bearer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401":
          description: "`invalid_credentials`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: "`too_many_attempts` (por e-mail e por IP)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/logout:
    post:
      summary: Encerra a sessão do token (provedor interno)
      operationId: authLogout
      tags: [auth]
      security: [{ bearer: [] }]
      responses:
        "204": { description: Sessão encerrada }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /v1/auth/email/verify:
    post:
      summary: Confirma o e-mail com o token do link enviado (provedor interno)
      operationId: authVerifyEmail
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthToken" }
      responses:
        "204": { description: E-mail confirmado }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "410":
          description: "`invalid_auth_token`: o link expirou ou já foi usado"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/email/resend:
    post:
      summary: Manda de novo o link de confirmação do e-mail (provedor interno)
      operationId: authResendVerification
      tags: [auth]
      security: [{ bearer: [] }]
      responses:
        "204": { description: "Link enviado, ou e-mail já confirmado" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "429":
          description: "`too_many_attempts`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/password/forgot:
    post:
      summary: Pede o link para redefinir a senha (provedor interno)
      description: >
        Responde 202 mesmo quando o e-mail não tem conta, para não revelar
        quem está cadastrado.
      operationId: authForgotPassword
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthForgotPassword" }
      responses:
        "202": { description: Pedido recebido }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "429":
          description: "`too_many_attempts`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/password/reset:
    post:
      summary: Define uma nova senha com o token do link enviado (provedor interno)
      description: Encerra todas as sessões abertas da conta.
      operationId: authResetPassword
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthResetPassword" }
      responses:
        "204": { description: Senha trocada }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "410":
          description: "`invalid_auth_token`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`weak_password`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # --- accounts ------------------------------------------------------------
...
  /v1/webhooks/billing:
    post:
      summary: Aviso de cobrança do gateway (webhook)
      description: >
        Recebe os avisos do gateway de cobrança (Asaas em produção). Não usa o
        token do app: a autenticidade vem do segredo do provedor, no cabeçalho
        que ele manda (`asaas-access-token` no Asaas). Um pagamento confirmado
        estende o acesso do workspace; um estorno o suspende na hora.
        Reenvios do mesmo aviso, eventos ignorados e avisos de assinaturas
        desconhecidas respondem 204, para o gateway não reenviar. Corpo de
        até 1 MB.
      operationId: billingWebhook
      tags: [billing]
      security: []
      parameters:
        - name: asaas-access-token
          in: header
          required: false
          description: Segredo do webhook configurado no painel do Asaas (gateway em produção).
          schema: { type: string }
      requestBody:
        required: true
        description: Corpo no formato do provedor.
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "204": { description: Aviso processado (ou sem efeito) }
        "401":
          description: "`invalid_billing_webhook`: o aviso não veio do gateway, ou o corpo não foi reconhecido"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502":
          description: "`billing_unavailable`: o servidor está sem gateway de cobrança"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/workspaces/{workspaceId}/subscription:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      summary: A assinatura do workspace (dono e mentor)
      description: >
        Plano, preço, assentos e situação do acesso. Continua acessível com o
        workspace suspenso: é por aqui que ele volta.
      operationId: getSubscription
      tags: [billing]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Assinatura
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403":
          description: "`billing_managers_only`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
    post:
      summary: Contratar o plano (só o dono)
      description: >
        Abre a assinatura mensal no gateway e devolve a fatura em aberto, que
        quem paga quita por PIX, boleto ou cartão. O acesso só é liberado
        quando o pagamento é confirmado.
      operationId: subscribe
      tags: [billing]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NewSubscription" }
      responses:
        "201":
          description: Assinatura criada, aguardando pagamento
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403":
          description: "`billing_owner_only`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
        "409":
          description: "`already_subscribed`, `seats_in_use` ou `no_price`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`invalid_tax_id`, `minimum_seats` ou `seats_above_plan`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: "`billing_unavailable`: o gateway não respondeu"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    patch:
      summary: Mudar os assentos da mentoria (só o dono)
      description: >
        Diminuir vale na hora, e nunca abaixo dos assentos em uso (afiliados
        mais convites pendentes). Aumentar vale quando o próximo pagamento for
        confirmado; a cobrança em aberto já sai com o novo valor.
      operationId: changeSeats
      tags: [billing]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeSeats" }
      responses:
        "200":
          description: Assinatura atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403":
          description: "`billing_owner_only`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: "`no_subscription` ou `workspace_not_found`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: "`seats_in_use` ou `no_price`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`minimum_seats` ou `seats_above_plan`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: "`billing_unavailable`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: Cancelar a assinatura (só o dono)
      description: >
        Encerra a cobrança no gateway. O acesso continua até o fim do período
        já pago.
      operationId: cancelSubscription
      tags: [billing]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Assinatura cancelada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403":
          description: "`billing_owner_only`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: "`no_subscription` ou `workspace_not_found`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: "`billing_unavailable`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/workspaces/{workspaceId}/subscription/simulate-payment:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      summary: Simular o pagamento (só no ambiente local)
      description: >
        Confirma a cobrança em aberto sem gateway de verdade. Existe apenas
        quando a cobrança está em modo mock (`simulated` na assinatura); fora
        dele devolve 404 `simulation_unavailable`.
      operationId: simulatePayment
      tags: [billing]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Pagamento aplicado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403":
          description: "`billing_owner_only`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: "`simulation_unavailable`, `no_subscription` ou `workspace_not_found`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: >
        Token de acesso do provedor de identidade: JWT do Zitadel (OIDC),
        token opaco do provedor interno (`/v1/auth/login`) ou `dev:<sub>` no
        modo de desenvolvimento.
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthenticated:
      description: Sem token ou token inválido (`unauthenticated`)
      headers:
        WWW-Authenticate:
          schema: { type: string, const: Bearer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidJSON:
      description: Corpo que não é JSON, com campo desconhecido ou grande demais (`invalid_json`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    WorkspaceNotFound:
      description: "`workspace_not_found`: o workspace não existe ou o usuário não é membro dele"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: "`internal_error`"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >
            Código estável para o front. Genéricos: unauthenticated,
            internal_error, invalid_json, invalid_request. Auth (provedor
            interno): invalid_credentials, email_taken, invalid_email,
            invalid_name, weak_password, invalid_auth_token,
            too_many_attempts. Contas: workspace_not_found, member_not_found,
            invite_not_found, forbidden, mentorship_only,
            consent_mentorship_only, owner_cannot_leave, no_seats,
            seats_in_use, seats_above_plan, workspace_suspended,
            workspace_suspended_owner, already_member, invite_expired,
            invite_used, invite_revoked, invite_other_email,
            invalid_workspace_name, invalid_photo_url, invalid_invite_email,
            invalid_invite_validity. Assinatura: billing_owner_only,
            billing_managers_only, already_subscribed, no_subscription,
            invalid_tax_id, minimum_seats, no_price, simulation_unavailable,
            billing_unavailable, invalid_billing_webhook. Shopee:
            invalid_shopee_credential, shopee_access_denied,
            shopee_rate_limited, shopee_unavailable, invalid_app_id,
            invalid_secret. Catálogo e radar: product_not_found,
            invalid_page, invalid_per_page, negative_price,
            invalid_min_commission, invalid_min_rating, invalid_sort,
            query_too_long, invalid_history_days, invalid_product_link,
            short_link. Coleções: item_not_found, collection_not_found,
            collection_exists, no_credential, import_unavailable,
            link_rate_limited, product_lookup_failed,
            product_or_link_required, invalid_item_title,
            invalid_item_description, invalid_item_notes, invalid_tag,
            too_many_tags, invalid_item_status, invalid_affiliate_link,
            invalid_collection_name, too_many_item_collections,
            invalid_item_page. Notificações: notification_not_found,
            push_unavailable, invalid_push_subscription,
            too_many_push_subscriptions, invalid_notification_preferences.
            Mídia: video_not_found, invalid_video_link, video_unavailable,
            platform_unavailable, usage_rights_required,
            invalid_video_format, invalid_video_size, video_too_large,
            video_quota_exceeded, uploads_unavailable, upload_closed,
            upload_incomplete, upload_parts_required, invalid_part_number,
            video_not_ready, video_owner_only, video_share_managers_only,
            video_title_too_long, embed_not_downloadable. Curadoria:
            list_not_found, list_item_not_found, video_not_in_list,
            list_managers_only, lists_mentorship_only, list_limit,
            list_empty, already_in_list, list_full, list_not_published,
            invalid_list_title, invalid_list_description,
            invalid_list_comment, list_comment_required,
            invalid_list_order, product_not_in_list. Resultados:
            group_results_forbidden, group_results_mentorship_only,
            sync_no_credential, synced_recently, invalid_period,
            sync_credential_refused, sync_shopee_unavailable.
        message: { type: string, description: Texto em português pronto para mostrar ao usuário }
...
    AuthRegister:
      type: object
      required: [name, email, password]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
    AuthLogin:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthToken:
      type: object
      required: [token]
      additionalProperties: false
      properties:
        token: { type: string }
    AuthForgotPassword:
      type: object
      required: [email]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
    AuthResetPassword:
      type: object
      required: [token, password]
      additionalProperties: false
      properties:
        token: { type: string }
        password: { type: string, minLength: 8, maxLength: 128 }
    AuthSession:
      type: object
      required: [token, expires_at]
      properties:
        token:
          type: string
          description: Token opaco da sessão, enviado como Bearer.
        expires_at:
          type: string
          format: date-time
          description: A validade é renovada a cada uso.

    # accounts
...
    Plan:
      type: string
      enum: [solo, mentorship]
    AccessStatus:
      type: string
      enum: [trial, active, free, suspended]
      description: >
        `trial` no período de avaliação, `active` com a assinatura paga,
        `free` no workspace pessoal de quem é afiliado de uma mentoria em dia
        (não é cobrado, por enquanto) e `suspended` quando o acesso venceu.
        Suspenso, o workspace só permite ver a si mesmo, sair dele (ou
        remover afiliados) e cuidar da assinatura; o resto devolve 402.
...
    SubscriptionStatus:
      type: string
      enum: [none, pending, active, overdue, cancelled]
    Subscription:
      type: object
      required:
        [plan, access_status, access_until, status, provider, price_cents,
         seats, seats_in_use, max_seats, amount_cents, next_due_date,
         payment_url, simulated]
      properties:
        plan: { $ref: "#/components/schemas/Plan" }
        access_status: { $ref: "#/components/schemas/AccessStatus" }
        access_until:
          type: string
          format: date-time
          description: Até quando o workspace pode ser usado (fim do teste ou do ciclo pago, mais a tolerância).
        status: { $ref: "#/components/schemas/SubscriptionStatus" }
        provider: { type: string, description: "Gateway de cobrança (ex.: asaas); vazio sem assinatura." }
        price_cents:
          type: integer
          format: int64
          description: >
            Preço mensal do plano: do workspace no solo, de cada assento na
            mentoria.
        seats: { type: integer, format: int64, description: Assentos contratados (0 sem assinatura). }
        seats_in_use: { type: integer, format: int64, description: Afiliados mais convites pendentes. }
        max_seats: { type: integer, format: int64, description: Teto do plano. }
        amount_cents: { type: integer, format: int64, description: Total mensal da assinatura (0 sem assinatura). }
        next_due_date:
          type: [string, "null"]
          format: date
          description: Vencimento da cobrança em aberto (ou da próxima), AAAA-MM-DD.
        payment_url:
          type: [string, "null"]
          description: Fatura em aberto no gateway (PIX, boleto ou cartão).
        simulated:
          type: boolean
          description: >
            Ambiente local com cobrança em modo mock: a tela pode oferecer
            simular o pagamento.
    NewSubscription:
      type: object
      required: [tax_id]
      additionalProperties: false
      properties:
        seats:
          type: integer
          format: int64
          minimum: 1
          description: Assentos da mentoria. Ignorado no plano solo.
        tax_id: { type: string, description: "CPF ou CNPJ de quem paga, com ou sem pontuação, exigido pelo gateway." }
    ChangeSeats:
      type: object
      required: [seats]
      additionalProperties: false
      properties:
        seats: { type: integer, format: int64, minimum: 1 }